requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==22.1.0
bcrypt==4.1.2
black==23.7.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.6
cryptography==42.0.5
distlib==0.3.7
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.0.4
fastapi==0.110.0
filelock==3.12.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.4
iniconfig==1.1.1
isort==5.12.0
JsonForm==0.0.2
jsonschema==4.17.0
JsonSir==0.0.2
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.25.2
packaging==22.0
passlib==1.7.4
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.9.1
pluggy==1.0.0
psutil==5.9.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.2
pydantic_core==2.10.1
Pympler==1.0.1
pyperclip==1.8.2
pyrsistent==0.19.2
pytest==7.2.0
python-dotenv==1.0.1
Python-EasyConfig==0.1.7
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.6
PyYAML==6.0
requests==2.31.0
Resource==0.2.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
urllib3==1.26.12
uvicorn==0.27.1
virtualenv==20.24.1
watchfiles==0.21.0
websockets==12.0
Whoosh==2.7.4
wincertstore==0.2
